🎉 feat(Cargo.toml): Add Chapter 3 as a member of the workspace
This commit adds "Ch.3 Writing Web Crawlers" as a member of the workspace in the Cargo.toml file. Now, Chapter 3 is included in the project workspace.

✨ feat(Cargo.toml): Create Chapter 3's Cargo.toml file

This commit adds the Cargo.toml file for "Ch.3 Writing Web Crawlers" chapter. It specifies dependencies required for this chapter's code.

📝 docs(Ch.3 Writing Web Crawlers): Add Readme.md

This commit adds an initial Readme.md file for "Ch.3 Writing Web Crawlers." The Readme.md is currently empty and can be filled with relevant content.

🎉 feat(Ch.3 Writing Web Crawlers): Add initial main.rs

This commit adds an initial main.rs file to "Ch.3 Writing Web Crawlers." The file currently contains a basic "Hello world!" program and can be further developed for web crawling purposes.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>