🔒 chore(Ch.5): Update database credentials handling ✨ feat(Ch.5): Integrate dotenv for environment variable management
In order to enhance security and streamline environment variable usage, this commit introduces changes to how database credentials are handled in the codebase. Additionally, the integration of the `dotenvy` crate is implemented to manage environment variables more efficiently.

Changes made:

- Import the necessary `dotenv` and `std::env` modules to facilitate environment variable management.
- The `dotenv().ok()` function is called to load environment variables from a `.env` file if available.
- Database user and password are retrieved from environment variables using the `env::var` function, providing improved security over hardcoding sensitive information.
- The `opts` variable, used for configuring the MySQL connection, is updated to use the retrieved database user and password.

By employing these changes, the commit ensures that sensitive credentials are not hard-coded within the source code, thereby enhancing security and maintainability.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>