🐛 fix(Ch.5 Storing Data): Fix import statements and error type naming
The import statements and error type naming have been updated to follow the recommended conventions. Additionally, the `EmailSender` and `WebScraper` traits now have async methods, and a new struct `SmtpEmailSender` has been introduced to implement the `EmailSender` trait for `SmtpTransport`. The logic for checking if it's Christmas and sending an email has been encapsulated in the `ChristmasChecker` struct.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>